pyproject.toml
[tool.poetry]
name = "poker"
version = "1.0a"
description = "Backend for playing poker via websockets."
authors = ["raindinners"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
colorlog = "^6.8.2"
pokerengine = "^1.6.1"
apscheduler = "^3.10.4"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
fastapi = "^0.103.1"
fastapi-distributed-websocket = "^0.2.0"
uvicorn = "^0.23.2"
redis = "^5.0.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
black = "^23.3.0"
ruff = "^0.0.292"
isort = "^5.12.0"
mypy = "^1.2.0"

[tool.ruff]
line-length = 99
select = [
    "C4",
    "E",
    "F",
    "T10",
    "T20",
    "Q",
    "RET",
]
ignore = [
    "E501",
]
exclude = [
    ".git",
    ".venv",
    "logs",
]
force-exclude = true

[tool.black]
line-length = 99
target-version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
    \.git
    \.ruff_cache
    \.venv
  | logs
)
'''

[tool.isort]
profile = "black"
src_paths = [
    "app",
    "tests",
]

[tool.mypy]
plugins = [
    "pydantic.mypy",
]
disable_error_code = [
    "import",
]
show_error_codes = true
show_error_context = true
namespace_packages = true

[[tool.mypy.overrides]]
module = [
    "passlib",
]
ignore_missing_imports = true
disallow_untyped_defs = true
